requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-froala-editor==4.0.9
idna==3.3
numpy==1.22.2
pandas==1.4.1
pdfkit==1.0.0
Pillow==9.0.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
